Fix issue #65
When filling cache by voltage level id (if cache is not fullyLoaded), data previously retained in cache is kept and merged with data newly retrieved from database. This ensures that data created in cache before cache is fully loaded is not removed from cache even if it is not present in database.

Signed-off-by: NOIR Nicolas ext <nicolas.noir@rte-france.com>