use spring's configured resttemplatebuilder instead of recreating one (#569)
The one from spring benefits from autoconfiguration (e.g. ObservationRestTemplateCustomizer
to automatically propagate micrometer tracing accross resttemplate calls)
Another huge benefit is to allow to use spring.http.client.factory to choose the resttemplate impl, with spring boot 3.4 without this configuration it autodetects from the classpath. In our case, on servers using spring-cloud-stream which brings reactor in the classpath, reactor is used and it causes hangs in our servers we don't know why but we want to avoid it.

Also add one small test for prod configuration (with prefix)

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>